+++ tests/test_file_data.py
@@ -2,9 +2,6 @@ from dataclasses import dataclass
 from datetime import date
 
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
-from utils.column_names.raw_data_files.cqc_location_api_columns import (
-    NewCqcLocationApiColumns as CQCL,
-)
 from utils.column_values.categorical_column_values import (
     AscwdsFilteringRule,
     CareHome,
@@ -17,368 +14,6 @@ from utils.raw_data_adjustments import RecordsToRemoveInLocationsData
 from utils.validation.validation_rule_custom_type import CustomValidationRules
 from utils.validation.validation_rule_names import RuleNames as RuleName
 from utils.column_values.categorical_column_values import MainJobRoleLabels
-
-
-@dataclass
-class ExtractRegisteredManagerNamesData:
-    extract_registered_manager_rows = [
-        (
-            "1-001",
-            date(2024, 1, 1),
-            [
-                {
-                    CQCL.name: "Activity 1",
-                    CQCL.code: "A1",
-                    CQCL.contacts: [
-                        {
-                            CQCL.person_family_name: "Surname_1a",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                        {
-                            CQCL.person_family_name: "Surname_1b",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                    ],
-                },
-            ],
-        ),
-        (
-            "1-002",
-            date(2024, 1, 1),
-            [
-                {CQCL.name: "Activity 2a", CQCL.code: "A2a", CQCL.contacts: []},
-                {
-                    CQCL.name: "Activity 2b",
-                    CQCL.code: "A2b",
-                    CQCL.contacts: [
-                        {
-                            CQCL.person_family_name: "Surname_2b",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                    ],
-                },
-            ],
-        ),
-    ]
-
-    extract_contacts_information_two_activities_one_contact_each_rows = [
-        (
-            "1-002",
-            date(2024, 1, 1),
-            [
-                {
-                    CQCL.name: "Activity 2a",
-                    CQCL.code: "A2a",
-                    CQCL.contacts: [
-                        {
-                            CQCL.person_family_name: "Surname_2a",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                    ],
-                },
-                {
-                    CQCL.name: "Activity 2b",
-                    CQCL.code: "A2b",
-                    CQCL.contacts: [
-                        {
-                            CQCL.person_family_name: "Surname_2b",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                    ],
-                },
-            ],
-        ),
-    ]
-    expected_extract_contacts_information_two_activities_one_contact_each_rows = [
-        (
-            "1-002",
-            date(2024, 1, 1),
-            [
-                {
-                    CQCL.name: "Activity 2a",
-                    CQCL.code: "A2a",
-                    CQCL.contacts: [
-                        {
-                            CQCL.person_family_name: "Surname_2a",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                    ],
-                },
-                {
-                    CQCL.name: "Activity 2b",
-                    CQCL.code: "A2b",
-                    CQCL.contacts: [
-                        {
-                            CQCL.person_family_name: "Surname_2b",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                    ],
-                },
-            ],
-            {
-                CQCL.person_family_name: "Surname_2a",
-                CQCL.person_given_name: "Name",
-                CQCL.person_roles: ["Registered Manager"],
-                CQCL.person_title: "M",
-            },
-        ),
-        (
-            "1-002",
-            date(2024, 1, 1),
-            [
-                {
-                    CQCL.name: "Activity 2a",
-                    CQCL.code: "A2a",
-                    CQCL.contacts: [
-                        {
-                            CQCL.person_family_name: "Surname_2a",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                    ],
-                },
-                {
-                    CQCL.name: "Activity 2b",
-                    CQCL.code: "A2b",
-                    CQCL.contacts: [
-                        {
-                            CQCL.person_family_name: "Surname_2b",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                    ],
-                },
-            ],
-            {
-                CQCL.person_family_name: "Surname_2b",
-                CQCL.person_given_name: "Name",
-                CQCL.person_roles: ["Registered Manager"],
-                CQCL.person_title: "M",
-            },
-        ),
-    ]
-
-    extract_contacts_information_one_activity_two_contacts_rows = [
-        (
-            "1-001",
-            date(2024, 1, 1),
-            [
-                {
-                    CQCL.name: "Activity 1",
-                    CQCL.code: "A1",
-                    CQCL.contacts: [
-                        {
-                            CQCL.person_family_name: "Surname_1a",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                        {
-                            CQCL.person_family_name: "Surname_1b",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                    ],
-                },
-            ],
-        )
-    ]
-    expected_extract_contacts_information_one_activity_two_contacts_rows = [
-        (
-            "1-001",
-            date(2024, 1, 1),
-            [
-                {
-                    CQCL.name: "Activity 1",
-                    CQCL.code: "A1",
-                    CQCL.contacts: [
-                        {
-                            CQCL.person_family_name: "Surname_1a",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                        {
-                            CQCL.person_family_name: "Surname_1b",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                    ],
-                },
-            ],
-            {
-                CQCL.person_family_name: "Surname_1a",
-                CQCL.person_given_name: "Name",
-                CQCL.person_roles: ["Registered Manager"],
-                CQCL.person_title: "M",
-            },
-        ),
-        (
-            "1-001",
-            date(2024, 1, 1),
-            [
-                {
-                    CQCL.name: "Activity 1",
-                    CQCL.code: "A1",
-                    CQCL.contacts: [
-                        {
-                            CQCL.person_family_name: "Surname_1a",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                        {
-                            CQCL.person_family_name: "Surname_1b",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                    ],
-                },
-            ],
-            {
-                CQCL.person_family_name: "Surname_1b",
-                CQCL.person_given_name: "Name",
-                CQCL.person_roles: ["Registered Manager"],
-                CQCL.person_title: "M",
-            },
-        ),
-    ]
-
-    extract_contacts_information_two_activities_but_only_one_with_contacts_rows = [
-        (
-            "1-002",
-            date(2024, 1, 1),
-            [
-                {CQCL.name: "Activity 2a", CQCL.code: "A2a", CQCL.contacts: []},
-                {
-                    CQCL.name: "Activity 2b",
-                    CQCL.code: "A2b",
-                    CQCL.contacts: [
-                        {
-                            CQCL.person_family_name: "Surname_2b",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                    ],
-                },
-            ],
-        ),
-    ]
-    expected_extract_contacts_information_two_activities_but_only_one_with_contacts_rows = [
-        (
-            "1-002",
-            date(2024, 1, 1),
-            [
-                {CQCL.name: "Activity 2a", CQCL.code: "A2a", CQCL.contacts: []},
-                {
-                    CQCL.name: "Activity 2b",
-                    CQCL.code: "A2b",
-                    CQCL.contacts: [
-                        {
-                            CQCL.person_family_name: "Surname_2b",
-                            CQCL.person_given_name: "Name",
-                            CQCL.person_roles: ["Registered Manager"],
-                            CQCL.person_title: "M",
-                        },
-                    ],
-                },
-            ],
-            {
-                CQCL.person_family_name: "Surname_2b",
-                CQCL.person_given_name: "Name",
-                CQCL.person_roles: ["Registered Manager"],
-                CQCL.person_title: "M",
-            },
-        ),
-    ]
-
-    select_and_create_full_name_rows = [
-        (
-            "1-001",
-            date(2024, 1, 1),
-            "Y",
-            {
-                CQCL.person_family_name: "Surname_1",
-                CQCL.person_given_name: "Name",
-                CQCL.person_roles: ["Registered Manager"],
-                CQCL.person_title: "M",
-            },
-        ),
-        (
-            "1-002",
-            date(2024, 1, 1),
-            "Y",
-            {
-                CQCL.person_family_name: "Surname_2",
-                CQCL.person_given_name: "Name",
-                CQCL.person_roles: ["Registered Manager"],
-                CQCL.person_title: "M",
-            },
-        ),
-    ]
-    expected_select_and_create_full_name_rows = [
-        ("1-001", date(2024, 1, 1), "Name Surname_1"),
-        ("1-002", date(2024, 1, 1), "Name Surname_2"),
-    ]
-
-    group_and_collect_names_with_duplicate_names_rows = [
-        ("1-001", date(2024, 1, 1), "Y", "Name Surname_1"),
-        ("1-001", date(2024, 1, 1), "Y", "Name Surname_1"),
-        ("1-001", date(2024, 1, 1), "Y", "Name Surname_2"),
-    ]
-    expected_group_and_collect_names_with_duplicate_names_rows = [
-        ("1-001", date(2024, 1, 1), ["Name Surname_2", "Name Surname_1"]),
-    ]
-
-    group_and_collect_names_with_different_ids_and_dates_rows = [
-        ("1-001", date(2024, 1, 1), "Y", "Name Surname_1"),
-        ("1-001", date(2024, 2, 1), "Y", "Name Surname_1"),
-        ("1-001", date(2024, 2, 1), "Y", "Name Surname_2"),
-        ("1-002", date(2024, 1, 1), "N", "Name Surname_3"),
-    ]
-    expected_group_and_collect_names_with_different_ids_and_dates_rows = [
-        ("1-001", date(2024, 1, 1), ["Name Surname_1"]),
-        ("1-001", date(2024, 2, 1), ["Name Surname_2", "Name Surname_1"]),
-        ("1-002", date(2024, 1, 1), ["Name Surname_3"]),
-    ]
-
-    original_test_rows = [
-        ("1-001", date(2024, 1, 1), "Y", 5),
-        ("1-001", date(2024, 2, 1), "Y", 5),
-        ("1-002", date(2024, 1, 1), "Y", 10),
-        ("1-002", date(2024, 2, 1), "Y", 10),
-    ]
-    registered_manager_names_rows = [
-        ("1-001", date(2024, 1, 1), ["Name Surname_1"]),
-        ("1-001", date(2024, 2, 1), ["Name Surname_1", "Name Surname_2"]),
-        ("1-002", date(2024, 2, 1), ["Name Surname_3"]),
-    ]
-    expected_join_with_original_rows = [
-        ("1-001", date(2024, 1, 1), "Y", 5, ["Name Surname_1"]),
-        ("1-001", date(2024, 2, 1), "Y", 5, ["Name Surname_1", "Name Surname_2"]),
-        ("1-002", date(2024, 1, 1), "Y", 10, None),
-        ("1-002", date(2024, 2, 1), "Y", 10, ["Name Surname_3"]),
-    ]
 
 
 @dataclass
