+++ tests/test_file_schemas.py
@@ -40,220 +40,6 @@ from utils.column_names.validation_table_columns import Validation
 
 
 @dataclass
-class ExtractRegisteredManagerNamesSchema:
-    extract_registered_manager_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(
-                CQCLClean.imputed_regulated_activities,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.code, StringType(), True),
-                            StructField(
-                                CQCL.contacts,
-                                ArrayType(
-                                    StructType(
-                                        [
-                                            StructField(
-                                                CQCL.person_family_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_given_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_roles,
-                                                ArrayType(StringType(), True),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_title, StringType(), True
-                                            ),
-                                        ]
-                                    ),
-                                    True,
-                                ),
-                                True,
-                            ),
-                        ]
-                    ),
-                    True,
-                ),
-                True,
-            ),
-        ]
-    )
-
-    extract_contacts_information_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(
-                CQCLClean.imputed_regulated_activities,
-                ArrayType(
-                    StructType(
-                        [
-                            StructField(CQCL.name, StringType(), True),
-                            StructField(CQCL.code, StringType(), True),
-                            StructField(
-                                CQCL.contacts,
-                                ArrayType(
-                                    StructType(
-                                        [
-                                            StructField(
-                                                CQCL.person_family_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_given_name,
-                                                StringType(),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_roles,
-                                                ArrayType(StringType(), True),
-                                                True,
-                                            ),
-                                            StructField(
-                                                CQCL.person_title, StringType(), True
-                                            ),
-                                        ]
-                                    ),
-                                    True,
-                                ),
-                                True,
-                            ),
-                        ]
-                    ),
-                    True,
-                ),
-                True,
-            ),
-        ]
-    )
-    expected_extract_contacts_information_schema = StructType(
-        [
-            *extract_contacts_information_schema,
-            StructField(
-                CQCLClean.contacts_exploded,
-                StructType(
-                    [
-                        StructField(
-                            CQCL.person_family_name,
-                            StringType(),
-                            True,
-                        ),
-                        StructField(
-                            CQCL.person_given_name,
-                            StringType(),
-                            True,
-                        ),
-                        StructField(
-                            CQCL.person_roles,
-                            ArrayType(StringType(), True),
-                            True,
-                        ),
-                        StructField(CQCL.person_title, StringType(), True),
-                    ]
-                ),
-            ),
-        ]
-    )
-
-    select_and_create_full_name_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCL.care_home, StringType(), True),
-            StructField(
-                CQCLClean.contacts_exploded,
-                StructType(
-                    [
-                        StructField(
-                            CQCL.person_family_name,
-                            StringType(),
-                            True,
-                        ),
-                        StructField(
-                            CQCL.person_given_name,
-                            StringType(),
-                            True,
-                        ),
-                        StructField(
-                            CQCL.person_roles,
-                            ArrayType(StringType(), True),
-                            True,
-                        ),
-                        StructField(CQCL.person_title, StringType(), True),
-                    ]
-                ),
-            ),
-        ]
-    )
-    expected_select_and_create_full_name_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCLClean.contacts_full_name, StringType(), True),
-        ]
-    )
-
-    group_and_collect_names_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCL.care_home, StringType(), True),
-            StructField(CQCLClean.contacts_full_name, StringType(), True),
-        ]
-    )
-    expected_group_and_collect_names_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(
-                CQCLClean.registered_manager_names, ArrayType(StringType(), True), True
-            ),
-        ]
-    )
-
-    original_test_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCL.care_home, StringType(), True),
-            StructField(CQCL.number_of_beds, IntegerType(), True),
-        ]
-    )
-    registered_manager_names_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(
-                CQCLClean.registered_manager_names, ArrayType(StringType(), True), True
-            ),
-        ]
-    )
-    expected_join_with_original_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCLClean.cqc_location_import_date, DateType(), True),
-            StructField(CQCL.care_home, StringType(), True),
-            StructField(CQCL.number_of_beds, IntegerType(), True),
-            StructField(
-                CQCLClean.registered_manager_names, ArrayType(StringType(), True), True
-            ),
-        ]
-    )
-
-
-@dataclass
 class UtilsSchema:
     cqc_pir_schema = StructType(
         [
